vite server exits without tty set to true
This took me 2 days to track down. 

The server exits before running tests but only on CI, it works fine locally.

Wasn't sure what was causing it, but eventually found this line of code in the vite server.

https://github.com/vitejs/vite/blob/main/packages/vite/src/node/server/index.ts#L363

This caused my tests to seem to pass on gitlab ci, before running any tests. It basically exits with a 0 exit code as soon as it realises the ci doesn't expose a tty. I'm not sure any CI has a tty as it's not an interactive terminal.

So the one line hack above seems to fix this.

Hope this helps someone else. 

A more elegant solution maybe to raise an issue on vite, but they don't say anywhere you can use their server in a CI, so this feel like a better place to fix this.

Thanks for the work on this so far. Perhaps add your github url to npm, this repo took a bit of effort to find.